Use string tokenizer for attribute split
